+++ docs/CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 2.0.0
+
+### Changed
+
+* Use one dynamic component instead of 8 (this also allows any of the style props to be animated)
+* Handle event animations in addition to static animations
+* Update README
+
 ## 1.0.10
 
 ### Added